requirements.txt
# Core dependencies
twilio==9.0.4
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Data validation
phonenumbers==8.13.29
validators==0.22.0

# Development (optional)
black==23.12.1
flake8==7.0.0
pytest==7.4.4
